fix delete book, order management ui

+++ Application/app/src/main/java/com/example/basalasa/model/entity/Book.kt
@@ -17,7 +17,7 @@ class Book(
     val description: String,
     val quantity: Int,
     val state: Int,
-    val star: Int,
+    val star: Double,
     var comments:ArrayList<Comments>?=null,
     var relatedBook:ArrayList<Book>?=null
 ) {
